requirements.txt
Flask
tensorflow
numpy
pillow
flask-cors
google-generativeai
gunicorn
